// lab12/todo-front/src/app/shared/models/models.ts
export interface ITaskList {
    id: number;
    name: string;
}

export interface ITask {
    id: number;
    name: string;
    created_at: string;
    due_on: string;
    status: string;
}

export interface ITaskDetail {
    id: number;
    name: string;
    status: string;
}

export interface ITaskNew {
    id: number,
    name: string,
    created_at: string,
    due_on: string,
    status: string,
    task_list: ITaskList
}